rename : #9 user 디렉터리 밑에 enums 디렉터리 생성
